And another one down

+++ common/commonutil.h
@@ -77,7 +77,7 @@ uint32_t reflect32(uint32_t b);      // dedicated 32bit reversal
 uint64_t reflect64(uint64_t b);      // dedicated 64bit reversal
 
 void num_to_bytes(uint64_t n, size_t len, uint8_t *dest);
-uint64_t bytes_to_num(const uint8_t *src, size_t len);
+uint64_t bytes_to_num(const size_t *src, size_t len);
 
 // LE and BE to/from memory
 uint16_t MemLeToUint2byte(const uint8_t *data);
